Move the showFlashMessage function to the helpers.js file and export it
